requirements_reviewer1.txt
numpy
pandas
PyYAML
Pillow
tqdm
matplotlib
